% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_sirene.R
\name{validation_sirene}
\alias{validation_sirene}
\title{Validité de SIRET ou SIREN}
\usage{
validation_sirene(id, type = c("siret", "siren"), warn = TRUE)
}
\arguments{
\item{id}{un vecteur caractère de SIRET, ou de SIREN si \code{type = "siren"}.}

\item{type}{par défaut \code{"siret"} (chaîne de caractères de 14 chiffres).
Autre valeur possible \code{"siren"} (9 chiffres).}

\item{warn}{\code{FALSE} pour désactiver d'éventuels warnings.}
}
\value{
Un vecteur booléen de même longueur que \code{id}. Une valeur manquante
produira une valeur manquante en retour.
}
\description{
Vérifie la validité d'un vecteur de SIRET ou de SIREN.
}
\details{
Pour les SIREN et SIRET, la vérification est faite sur la base de
la \href{https://fr.wikipedia.org/wiki/Formule_de_Luhn}{formule de Luhn}.

Un SIREN est valide si sa somme de Luhn est un multipe de 10.

Un SIRET est valide si sa somme de Luhn est un multipe de 10 et si la somme
de Luhn de son SIREN (9 premiers chiffres) est un multiple de 10.

Les SIRET de \emph{La Poste} (SIREN 356000000) ne respectant pas cette règle font
l'objet d'un traitement différencié. Il seront considérés comme valides si
la somme des chiffres les composant est un multiple de 5.
}
\examples{
validation_sirene(c("20003452800014", "20003452800041", "a", NA)) # SIRET (14 chiffres)
validation_sirene(c("200034528", "200034582", "a", NA), type = "siren")
}
